# docs/swagger.yaml
basePath: /
...
host: localhost:8081
info:
  contact: {}
  description: |-
    RESTful API for a Hotel Booking System built with Go, Gin, GORM, PostgreSQL, and JWT authentication.

    Features include authentication, hotel management, room management, bookings, availability, reviews, favorites, and payments.
  title: Hotel Booking System API
  version: "1.0"
paths:
...
  /hotels/{id}/rating:
    get:
      description: Get the average rating of a specific hotel based on submitted reviews.
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get hotel average rating
      tags:
      - Reviews
  /hotels/{id}/reviews:
    get:
      description: Get all reviews submitted for a specific hotel.
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get hotel reviews
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Create a review and rating for a hotel. The authenticated user
        can submit one review per hotel.
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review data
        in: body
        name: request
        required: true
        schema:
          properties:
            comment:
              type: string
            rating:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a hotel review
      tags:
      - Reviews
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"